+++ src/Geometry.hs
@@ -20,12 +20,13 @@ where
 
 import Data.List (find)
 import Linear
-  ( Additive ((^+^)),
+  ( Additive (..),
     Epsilon (..),
-    Metric (dot),
+    Metric (..),
     Quaternion,
     V3 (..),
     axisAngle,
+    cross,
     normalize,
     (*^),
     (^*),
@@ -45,11 +46,9 @@ data Ray = Ray {origin :: Position, direction :: Direction} deriving (Eq, Show)
 
 data Hit = Hit {hitColour :: Colour, hitTime :: Float, hitPoint :: Position, hitNormal :: Direction} deriving (Eq, Show)
 
-data Shape = Sphere
-  { sphereCenter :: Position,
-    sphereRadius :: Float,
-    sphereColour :: Colour
-  }
+data Shape
+  = Sphere Colour Position Float
+  | Triangle Colour Position Position Position
   deriving (Show)
 
 solveQuadratic :: (Ord a, Epsilon a, Floating a) => a -> a -> a -> [a]
@@ -64,19 +63,26 @@ solveQuadratic a b c
 instance Ord Hit where
   (Hit {hitTime = t1}) `compare` (Hit {hitTime = t2}) = t1 `compare` t2
 
-intersect :: Shape -> Ray -> Maybe Hit
-intersect sphere ray =
-  hit <$> intersectSphere (sphereCenter sphere) (sphereRadius sphere) ray
+toHit :: Colour -> (Position -> Direction) -> Ray -> Float -> Hit
+toHit colour pointToNormal ray time =
+  Hit
+    { hitTime = time,
+      hitColour = colour,
+      hitPoint = rayHitPoint,
+      hitNormal = pointToNormal rayHitPoint
+    }
   where
-    hit time =
-      Hit
-        { hitTime = time,
-          hitColour = sphereColour sphere,
-          hitPoint = rayHitPoint,
-          hitNormal = normalize $ rayHitPoint - sphereCenter sphere
-        }
-      where
-        rayHitPoint = origin ray ^+^ (direction ray ^* time)
+    rayHitPoint = origin ray ^+^ (direction ray ^* time)
+
+intersect :: Shape -> Ray -> Maybe Hit
+intersect shape ray = toHit (colour shape) (pointToNormal shape) ray <$> intersect' shape ray
+  where
+    intersect' (Sphere _ center radius) = intersectSphere center radius
+    intersect' (Triangle _ v0 v1 v2) = intersectTriangle v0 v1 v2
+    colour (Sphere c _ _) = c
+    colour (Triangle c _ _ _) = c
+    pointToNormal (Sphere _ center _) hit = normalize (hit ^-^ center)
+    pointToNormal (Triangle _ v0 v1 v2) _ = normalize ((v1 - v0) `cross` (v2 - v0))
 
 intersectSphere :: Position -> Float -> Ray -> Maybe Float
 intersectSphere center radius ray = find (> 0) (solveQuadratic a b c)
@@ -85,6 +91,25 @@ intersectSphere center radius ray = find (> 0) (solveQuadratic a b c)
     a = direction ray `dot` direction ray
     b = 2 * oc `dot` direction ray
     c = oc `dot` oc - radius * radius
+
+intersectTriangle :: Position -> Position -> Position -> Ray -> Maybe Float
+intersectTriangle v0 v1 v2 ray
+  | nearZero det = Nothing
+  | u < 0 || u > 1 = Nothing
+  | v < 0 || u + v > 1 = Nothing
+  | t < 0 = Nothing
+  | otherwise = Just t
+  where
+    v0v1 = v1 - v0
+    v0v2 = v2 - v0
+    pvec = direction ray `cross` v0v2
+    det = v0v1 `dot` pvec
+    invDet = 1 / det
+    tvec = origin ray - v0
+    u = (tvec `dot` pvec) * invDet
+    qvec = tvec `cross` v0v1
+    v = (direction ray `dot` qvec) * invDet
+    t = (v0v2 `dot` qvec) * invDet
 
 rotatePitch :: Float -> Rotation
 rotatePitch = axisAngle (V3 1 0 0)
@@ -102,7 +127,7 @@ schurProduct :: Num a => V3 a -> V3 a -> V3 a
 schurProduct (V3 x1 y1 z1) (V3 x2 y2 z2) = V3 (x1 * x2) (y1 * y2) (z1 * z2)
 
 fromTo :: Position -> Position -> Ray
-fromTo from to = Ray {origin = from, direction = normalize $ to - from}
+fromTo from to = Ray {origin = from, direction = normalize $ to ^-^ from}
 
 extrude :: Position -> Direction -> Position
-extrude point normal = point ^+^ (0.01 *^ normal)
+extrude point dir = point ^+^ (0.0001 *^ dir)